Initial import
* Can create a floating component which gets pseudoanimated
* When the floating time is beyond the animation time, the
  floating component is removed so the system stops to run

In the future I think this will be replaced by a tag component
called IsAnimating and a separate system will just check world
time to figure out which floaters need to be added/removed from
the animating set.

The next set of entities will be the waypoints between which the
floaters will go.